Use a new HTTP client for each CI loop
This also updates send() to throw ClientException as required by
package:http, rather than an Error subclass.
